Refactor ancestral set functions to submodule
